major simplification of node structure
The previous development plan still compiles, but needs to be adapted.

Add int value to each node as data payload.
Get rid of unode and cnode, everything is a binary node (BNode) now.
Commands are recognized by magic number in the value field of the node.
unode and cnode are represented by a bnode with the left or both arrows pointing to itself. (You can no longer have a unode pointing to itself a:[a], this becomes a cnode)
Trimmer: Use self-reference instead of marker
Runner is more robust (for better and worse), as you always find a suitable structure for executing the next step.
Just the command may not be recognized from the value.

Change parser to insert nodes into the net from left to right (rather than depth first). This helps with more consistent formatting afterwards.

Add flag to force formatter to assign a given name, even if the node occurs only once.